requirements.txt

chromadb
fastapi
langchain
libcst
pathspec
pydantic
pydantic-settings
pygments
python-magic
sentence-transformers
tiktoken
torch
tqdm
transformers
tree-sitter
tree-sitter-languages
uvicorn
httpx
pytest-asyncio
slowapi
starlette

